added axios get push patch to various pages

+++ onestock_api/onestock_backend/models.py
@@ -51,7 +51,7 @@ class User(AbstractBaseUser, PermissionsMixin):
         return str(self.user_name)
     
 class Account(models.Model):
-    accountValue = models.IntegerField()
+    accountValue = models.FloatField()
     accountType = models.CharField(max_length=100)
     user = models.ForeignKey(User, on_delete = models.CASCADE)
     
@@ -60,16 +60,16 @@ class Account(models.Model):
 
 class Brokerage(models.Model):
     name = models.CharField(max_length=100)
-    fees = models.IntegerField()
-    perOrderFees = models.IntegerField()
-    currencyConversionRate = models.IntegerField()
+    fees = models.FloatField()
+    perOrderFees = models.FloatField()
+    currencyConversionRate = models.FloatField()
 
     def __str__(self) -> str:
         return self.name
 
 class OrderForm(models.Model):
     typeofFile = models.CharField(max_length = 100)
-    numOfOrders = models.IntegerField()
+    numOfOrders = models.FloatField()
     dateRequested = models.DateField(max_length = 100)
     account = models.ForeignKey(Account, on_delete=models.CASCADE)
 
@@ -79,8 +79,8 @@ class OrderForm(models.Model):
 class Order(models.Model):
     ticker = models.CharField(max_length = 100)
     buysell = models.CharField(max_length = 1)
-    quantity = models.IntegerField()
-    purchasePrice = models.IntegerField()
+    quantity = models.FloatField()
+    purchasePrice = models.FloatField()
     purchaseTime = models.DateTimeField(max_length = 100)
     orderReqID = models.ForeignKey(OrderForm, on_delete=models.CASCADE)
     
@@ -89,9 +89,9 @@ class Order(models.Model):
 
 class TotalStockHistory(models.Model):
     uniqueTickers = models.BigIntegerField()
-    totalInvested = models.BigIntegerField()
-    quantityOfTrades = models.BigIntegerField()
-    netValue = models.BigIntegerField()
+    totalInvested = models.FloatField()
+    quantityOfTrades = models.FloatField()
+    netValue = models.FloatField()
     user = models.ForeignKey(User, on_delete=CASCADE)
 
     def __str__(self) -> str:
@@ -100,10 +100,10 @@ class TotalStockHistory(models.Model):
 class Specific_Stock_History(models.Model):
     ticker = models.CharField(max_length=100, unique = True)
     industry = models.CharField(max_length = 100)
-    netProfit = models.IntegerField()
+    netProfit = models.FloatField()
     exchange = models.CharField(max_length = 100)
-    amountShares = models.IntegerField()
-    currentValue = models.IntegerField()
+    amountShares = models.FloatField()
+    currentValue = models.FloatField()
     stockHistoryID = models.ForeignKey(TotalStockHistory, on_delete=models.CASCADE, null=True, blank = True)
     user = models.ForeignKey(User, on_delete = models.CASCADE)
 
@@ -113,8 +113,8 @@ class Specific_Stock_History(models.Model):
 class Dividend (models.Model):
     ticker = models.ForeignKey(Specific_Stock_History, on_delete=CASCADE)
     date = models.DateField(unique=True)
-    Yield= models.IntegerField()
-    TaxRate = models.IntegerField()
+    Yield= models.FloatField()
+    TaxRate = models.FloatField()
     
     class Meta:
         unique_together = ('ticker', 'date')
@@ -123,7 +123,7 @@ class Dividend (models.Model):
         return self.ticker+": "+self.date
 
 class RealEstate(models.Model):
-    valuation = models.BigIntegerField()
+    valuation = models.FloatField()
     location = models.CharField(max_length = 100)
     type = models.CharField(max_length = 100)
     user = models.ForeignKey(User,on_delete=CASCADE)
@@ -132,7 +132,7 @@ class RealEstate(models.Model):
         return str(self.user)+": "+self.type
 
 class Commodities(models.Model):
-    valuation = models.BigIntegerField()
+    valuation = models.FloatField()
     type = models.CharField(max_length = 100)
     user = models.ForeignKey(User,on_delete=CASCADE)
 
@@ -140,27 +140,27 @@ class Commodities(models.Model):
         return str(self.user)+": "+self.type
 
 class CryptoCurrency(models.Model):
-    valuation = models.BigIntegerField()
+    valuation = models.FloatField()
     name = models.CharField(max_length = 100)
-    purchasePrice = models.BigIntegerField()
-    quantity = models.BigIntegerField()
+    purchasePrice = models.FloatField()
+    quantity = models.FloatField()
     user = models.ForeignKey(User,on_delete=CASCADE)
 
     def __str__(self) -> str:
         return str(self.user)+": "+self.name
 
 class Bonds(models.Model):
-    valuation = models.BigIntegerField()
+    valuation = models.FloatField()
     interest = models.CharField(max_length = 100)
     principal = models.CharField(max_length = 100)
-    maturityDate = models.BigIntegerField()
+    maturityDate = models.FloatField()
     user = models.ForeignKey(User,on_delete=CASCADE)
 
     def __str__(self) -> str:
         return str(self.user)+": "+self.valuation
 
 class Cash(models.Model):
-    valuation = models.BigIntegerField()
+    valuation = models.FloatField()
     bank = models.CharField(max_length = 100)
     currency = models.CharField(max_length = 100)
     user = models.ForeignKey(User,on_delete=CASCADE)
